Ingest Agent Configuration

+++ src/Aquarius.ONE.Ingest.Windows-Service/Agents/Test/TelemetryConfig.cs
@@ -8,8 +8,8 @@ namespace ONE.Ingest.WindowsService.Agents.Test
 {
     public class TelemetryConfig : IngestTelemetryConfiguration
     {
-        public decimal MinimumValue { get; set; }
-        public decimal MaximumValue { get; set; }
+        public double MinimumValue { get; set; }
+        public double MaximumValue { get; set; }
 
     }
 }
